Committing an untested new spec change system, uncomment #define UNTESTED_NEW_SPEC_CHANGE_SYSTEM to test.

+++ Trunk/src/hearthstone-world/SpellAuras.h
@@ -445,7 +445,7 @@ public:
 	void SetAuraSlot(uint8 slot) { m_auraSlot = slot; }
 	HEARTHSTONE_INLINE uint8 GetAuraFlags() const { return m_auraFlags; }
 	void SetAuraFlags(uint8 flags) { m_auraFlags = flags; }
-	void SetAuraFlag(uint16 flag) { if( ! (m_auraFlags & flag ) ) m_auraFlags |= flag; };
+	void SetAuraFlag(uint8 flag) { if( ! (m_auraFlags & flag ) ) m_auraFlags |= flag; };
 	HEARTHSTONE_INLINE uint8 GetAuraLevel() const { return m_auraLevel; }
 	void SetAuraLevel(uint8 level) { m_auraLevel = level; }
